Changed aesthetics, loading, top tracks
Changed some of the styles around. Added a loading screen that displays before everything is loaded. Improved some of the functionality of the Top Tracks (link of the playing song is now blue). Unfortunately IE11 seems to load tracks a bit slow, so I may also implement a loading icon for the top tracks.